+++ src/core/state/public.ts
@@ -1,2 +1,2 @@
 export { stateReconcile } from "./index.js";
-
+export { kickOpportunisticStateMaintenance, runOpportunisticStateMaintenanceWorker } from "./index.js";